'CommitMessage.txt'
- tweek label spacing for left and right labels (quick fix extra border EB = 5)
- remove top and bottom coertion for left and right borders
- quick fix: explicit convertion float to integer in layout.py:
    in _TileHorizontally() and _TileVertically() @ "for i in range(int(width/w)):"
- change label default font size: 8 to 10 pts.

+++ layout.py
@@ -418,7 +418,7 @@ class _decorationsLibrary():
 
     def _TileHorizontally(self, dc, tile, x, y, width):
         w, h = tile.GetSize()
-        for i in range(width / w):
+        for i in range(int(width/w)):
             dc.DrawBitmap(tile, x, y)
             x += w
         remainder = width % w
@@ -429,7 +429,7 @@ class _decorationsLibrary():
 
     def _TileVertically(self, dc, tile, x, y, height):
         w, h = tile.GetSize()
-        for i in range(height / h):
+        for i in range(int(height/h)):
             dc.DrawBitmap(tile, x, y)
             y += h
         remainder = height % h
